chapter_10/requirements.txt
openai
python-dotenv
playwright
